+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.5.1] - 2023-05-11
+
+### Added
+
+- `listeners` Method.
+- `listenersAny` Method.
+- `listenersAnyOutgoing` Method.
+- `off` Method.
+- `onAny` Method.
+- `onAnyOutgoing` Method.
+- `prependAny` Method.
+- `prependAnyOutgoing` Method.
+- `timeout` Method.
+- `volatile` Method.
+
 ## [4.5.0] - 2023-05-11
 
 ### Added